requirements.txt
# Core dependencies
watchdog>=3.0.0
python-frontmatter>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# MCP and async
mcp>=0.9.0
asyncio>=3.4.3
aiohttp>=3.9.1
httpx>=0.25.2

# Retry and circuit breaker
tenacity>=8.0.0
aiobreaker>=1.2.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.7

# Social Media APIs
requests>=2.31.0
requests-oauthlib>=1.3.1

# Email
aiosmtplib>=3.0.1

# Template rendering
jinja2>=3.1.2

# Date/time utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development tools
black>=23.12.1
flake8>=6.1.0
mypy>=1.7.1
isort>=5.13.2

# Optional: Redis for caching
redis>=5.0.1
aioredis>=2.0.1
